mac_version/requirements.txt
sherpa-onnx>=1.10.0
sounddevice>=0.4.6
pynput>=1.7.6
rumps>=0.4.0
pyobjc-framework-Cocoa>=10.0
pyobjc-framework-Quartz>=10.0
numpy>=1.24.0
requests>=2.31.0
